# internal/panel/openapi.yaml
openapi: 3.1.0
info:
  title: starport-panel API
  version: v1
  description: |
    自托管 Kubernetes 面板的公开 HTTP API。所有调用方（内嵌 Web UI、CLI、上层系统）一视同仁地使用这一套接口。

    - 前缀 `/api/v1`；响应一律 JSON（kubeconfig / 资源 YAML / 日志除外）。
    - 错误统一为 `{"error":{"code":"...","message":"..."}}`，调用方按 `code` 分支，不解析 `message`。
    - 鉴权：`Authorization: Bearer <token>`，或（仅 WebSocket）`?token=`。除 `POST /agents/register` 与 `/healthz` 外全部需要。
    - 异步操作（装机 / 脚本 / 移除 / 升级）返回 `202 {"taskId"}`，用 `/tasks/{id}/logs` 增量拉日志直到 `task.status != running`。
    - WebSocket 端点（终端 / exec）在此仅描述参数；帧协议见 docs/api.md「终端协议」。
    - 本文件随二进制提供：`GET /api/v1/openapi.yaml`（需鉴权）。
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0
servers:
  - url: /api/v1
security:
  - bearer: []
  - queryToken: []
...
paths:
...
  /nodes/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [nodes]
      summary: 单节点
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Node" } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [nodes]
      summary: 删除节点记录
      description: 只删面板记录并断开连接（其令牌随之失效）；节点仍是某集群成员时 `409 NODE_IN_CLUSTER`。机器上的 agent 服务需自行停掉，否则会重新注册成新节点。
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /nodes/{id}/exec:
    parameters: [{ $ref: "#/components/parameters/id" }]
    post:
      tags: [nodes]
      summary: 在节点上执行脚本（异步任务）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [script]
              properties:
                script: { type: string, description: 交给 `sh -c` 的脚本 }
                timeoutMs: { type: integer, format: int64, description: 0 = agent 默认 }
      responses:
        "202": { $ref: "#/components/responses/TaskAccepted" }
        "409": { $ref: "#/components/responses/Error", description: NODE_OFFLINE }
  /nodes/{id}/terminal:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [nodes]
      summary: 节点 Web 终端（WebSocket）
      description: 升级为 WebSocket；节点离线 `409 NODE_OFFLINE`。仅 Linux agent 支持 PTY。
      parameters:
        - { in: query, name: cols, schema: { type: integer, default: 120 } }
        - { in: query, name: rows, schema: { type: integer, default: 30 } }
        - { in: query, name: shell, schema: { type: string }, description: "`sh -c` 执行的脚本，默认 `exec bash -l`" }
        - { in: query, name: tty, schema: { type: boolean, default: true } }
      responses:
        "101": { description: WebSocket }
  /nodes/{id}/upgrade:
    parameters: [{ $ref: "#/components/parameters/id" }]
    post:
      tags: [nodes]
      summary: 升级该节点的 starport-agent（异步任务）
      description: 节点下载 `binUrl`、校验 `sha256`（可选）、试运行 `--version`、覆盖二进制，3 秒后重启 systemd 服务并以新版本重连。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpgradeRequest" }
      responses:
        "202": { $ref: "#/components/responses/TaskAccepted" }
        "409": { $ref: "#/components/responses/Error", description: NODE_OFFLINE }
  /nodes/upgrade:
    post:
      tags: [nodes]
      summary: 批量升级 agent
      description: "`nodeIds` 为空则对全部在线节点下发；离线节点进入 `skipped`。"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpgradeRequest" }
      responses:
        "202":
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: object
                    additionalProperties: { type: integer, format: int64 }
                    description: nodeId → taskId
                  skipped: { type: array, items: { type: integer, format: int64 } }

  # ── 集群 ─────────────────────────────────────────────────────────────────────
  /clusters:
    get:
      tags: [clusters]
      summary: 全部集群
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Cluster" } } } } }
    post:
      tags: [clusters]
      summary: 新建集群记录（尚无控制面）
      description: 之后经 `POST /clusters/{id}/nodes` 先加一台 `first-master` 才会真正装机。`artifactMode=bundle` 必须给 `bundleUrl`；`online` 只支持 `cni=calico`。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClusterRequest" }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Cluster" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error", description: CLUSTER_NAME_EXISTS }
  /clusters/probe:
    post:
      tags: [clusters]
      summary: 探测 kubeconfig（接管前预检，不入库）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kubeconfig]
              properties: { kubeconfig: { type: string } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ClusterProbe" } } } }
        "400": { $ref: "#/components/responses/Error", description: KUBECONFIG_UNREACHABLE }
  /clusters/import:
    post:
      tags: [clusters]
      summary: 接管已有集群
      description: 只需要 admin kubeconfig；先探测成功才入库，记录 `source=imported`、`status=ready`。这类集群只做资源管理，不能经面板加 / 移节点（`409 CLUSTER_IMPORTED`）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kubeconfig]
              properties:
                name: { type: string, pattern: "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$" }
                kubeconfig: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster: { $ref: "#/components/schemas/Cluster" }
                  probe: { $ref: "#/components/schemas/ClusterProbe" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error", description: CLUSTER_NAME_EXISTS }
  /clusters/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [clusters]
      summary: 集群详情（含成员）
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ClusterDetail" } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [clusters]
      summary: 删集群
      description: 仍有成员且未 `force` → `409 CLUSTER_HAS_MEMBERS`；`force=true` 对每个在线成员发 `kubeadm reset` 任务并立即删记录。接管的集群只删记录，不动集群本身。
      parameters:
        - { in: query, name: force, schema: { type: boolean, default: false } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties: { taskIds: { type: array, items: { type: integer, format: int64 } } }
        "409": { $ref: "#/components/responses/Error" }
  /clusters/{id}/nodes:
    parameters: [{ $ref: "#/components/parameters/id" }]
    post:
      tags: [clusters]
      summary: 把节点装进集群（异步任务）
      description: |
        顺序：先一台 `first-master`（集群 `created|failed` → `installing` → `ready`），之后才能 `join-master` / `worker`。
        一台机只能属于一个集群；同集群上次 `failed` 的成员允许以同角色重试。join 凭据过期时面板自动在在线 master 上重新签发。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nodeId, role]
              properties:
                nodeId: { type: integer, format: int64 }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "202": { $ref: "#/components/responses/TaskAccepted" }
        "409": { $ref: "#/components/responses/Error", description: NODE_OFFLINE / NODE_ALREADY_MEMBER / CLUSTER_NOT_READY / CLUSTER_HAS_CONTROL_PLANE / CLUSTER_IMPORTED }
  /clusters/{id}/nodes/{nodeId}:
    parameters:
      - { $ref: "#/components/parameters/id" }
      - { in: path, name: nodeId, required: true, schema: { type: integer, format: int64 } }
    delete:
      tags: [clusters]
      summary: 移除成员（异步任务）
      description: 在另一台在线 master 上 `drain` + `delete node`，再在该节点 `kubeadm reset`，成功后删成员。唯一控制面 `409 LAST_MASTER`；成员忙 `409 MEMBER_BUSY`；无在线 master `409 NO_ONLINE_MASTER`。
      responses:
        "202": { $ref: "#/components/responses/TaskAccepted" }
        "409": { $ref: "#/components/responses/Error" }
  /clusters/{id}/kubeconfig:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [clusters]
      summary: admin kubeconfig
      responses:
        "200": { content: { application/yaml: { schema: { type: string } } } }
        "409": { $ref: "#/components/responses/Error", description: CLUSTER_NOT_READY }
    put:
      tags: [clusters]
      summary: 替换 kubeconfig（接管集群证书轮换后）
      description: 保存前先探测一次；只允许 `source=imported` 的集群（`409 CLUSTER_NOT_IMPORTED`）。替换后面板丢弃该集群的客户端缓存。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kubeconfig]
              properties: { kubeconfig: { type: string } }
      responses:
        "204": { description: 已替换 }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }

  # ── 集群内 Kubernetes 资源 ──────────────────────────────────────────────────
  /clusters/{id}/k8s/nodes:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [k8s]
      summary: K8s 节点
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/K8sNode" } } } } }
        "409": { $ref: "#/components/responses/Error", description: CLUSTER_NOT_READY }
        "502": { $ref: "#/components/responses/Error", description: APISERVER_UNREACHABLE }
  /clusters/{id}/k8s/nodes/{name}/cordon:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/name" }]
    post:
      tags: [k8s]
      summary: 停止调度（cordon）
      responses: { "204": { description: ok } }
  /clusters/{id}/k8s/nodes/{name}/uncordon:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/name" }]
    post:
      tags: [k8s]
      summary: 恢复调度（uncordon）
      responses: { "204": { description: ok } }
...
  /clusters/{id}/k8s/namespaces:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [k8s]
      summary: 命名空间
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Namespace" } } } } }
  /clusters/{id}/k8s/pods:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/namespace" }]
    get:
      tags: [k8s]
      summary: Pod 列表
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Pod" } } } } }
  /clusters/{id}/k8s/namespaces/{ns}/pods/{name}:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    delete:
      tags: [k8s]
      summary: 删 Pod（由控制器重建）
      responses: { "204": { description: ok } }
  /clusters/{id}/k8s/namespaces/{ns}/pods/{name}/logs:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    get:
      tags: [k8s]
      summary: 容器日志
      parameters:
        - { in: query, name: container, schema: { type: string }, description: 多容器 Pod 必填 }
        - { in: query, name: tail, schema: { type: integer, default: 500 } }
        - { in: query, name: previous, schema: { type: boolean, default: false } }
        - { in: query, name: follow, schema: { type: boolean, default: false }, description: 分块流式直到断开或容器结束 }
      responses:
        "200": { content: { text/plain: { schema: { type: string } } } }
  /clusters/{id}/k8s/namespaces/{ns}/pods/{name}/exec:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    get:
      tags: [k8s]
      summary: 容器 Web 终端（WebSocket）
      parameters:
        - { in: query, name: container, schema: { type: string } }
        - { in: query, name: cmd, schema: { type: array, items: { type: string } }, description: 可重复给 argv；默认 bash，不存在回落 sh }
        - { in: query, name: cols, schema: { type: integer } }
        - { in: query, name: rows, schema: { type: integer } }
      responses:
        "101": { description: WebSocket }
  /clusters/{id}/k8s/deployments:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/namespace" }]
    get:
      tags: [k8s]
      summary: Deployment 列表
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Deployment" } } } } }
  /clusters/{id}/k8s/namespaces/{ns}/deployments/{name}/scale:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    post:
      tags: [k8s]
      summary: 扩缩
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replicas]
              properties: { replicas: { type: integer, minimum: 0 } }
      responses: { "204": { description: ok } }
  /clusters/{id}/k8s/namespaces/{ns}/deployments/{name}/restart:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    post:
      tags: [k8s]
      summary: 滚动重启（kubectl rollout restart）
      responses: { "204": { description: ok } }
  /clusters/{id}/k8s/namespaces/{ns}/deployments/{name}/expose:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    post:
      tags: [k8s]
      summary: 为 Deployment 建 / 更新 Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [port]
              properties:
                name: { type: string, description: 默认同 Deployment 名 }
                type: { type: string, enum: [ClusterIP, NodePort, LoadBalancer], default: ClusterIP }
                port: { type: integer }
                targetPort: { type: integer, description: 默认同 port }
                nodePort: { type: integer, description: 仅 NodePort；0 = 自动分配 }
                protocol: { type: string, enum: [TCP, UDP], default: TCP }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Service" } } } }
  /clusters/{id}/k8s/services:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/namespace" }]
    get:
      tags: [k8s]
      summary: Service 列表
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Service" } } } } }
  /clusters/{id}/k8s/namespaces/{ns}/services/{name}:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/ns" }, { $ref: "#/components/parameters/name" }]
    delete:
      tags: [k8s]
      summary: 删 Service
      responses: { "204": { description: ok } }
  /clusters/{id}/k8s/ingresses:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/namespace" }]
    get:
      tags: [k8s]
      summary: Ingress 列表
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Ingress" } } } } }
  /clusters/{id}/k8s/events:
    parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/namespace" }]
    get:
      tags: [k8s]
      summary: 事件（最近发生倒序）
      parameters:
        - { in: query, name: object, schema: { type: string }, description: "`Kind/name`，只看该对象" }
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/K8sEvent" } } } } }
  /clusters/{id}/k8s/apply:
    parameters: [{ $ref: "#/components/parameters/id" }]
    post:
      tags: [k8s]
      summary: server-side apply 多文档 YAML
      description: 等价 `kubectl apply --server-side --force-conflicts`（字段管理者 `starport-panel`）。无 `metadata.namespace` 的 namespaced 资源落到 `namespace` 参数（默认 default）。任一文档失败即停止，响应同时带已成功的 `applied` 与 `error`，已成功的不回滚。
      parameters:
        - { in: query, name: namespace, schema: { type: string, default: default } }
      requestBody:
        required: true
        content:
          application/yaml: { schema: { type: string } }
          application/json:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest: { type: string }
                namespace: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Applied" }
...
  /clusters/{id}/k8s/resource-kinds:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [k8s]
      summary: 集群支持的全部可 list 资源类型（含 CRD）
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ResourceKind" } } } } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 库内令牌（`starport-panel token create`）或静态令牌（`serve --api-token`）
    queryToken:
      type: apiKey
      in: query
      name: token
      description: 仅供浏览器 WebSocket（无法自定义 Header）
  parameters:
    id: { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
...
  responses:
    Error:
      description: 结构化错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Node:
      type: object
      properties:
        id: { type: integer, format: int64 }
        facts: { $ref: "#/components/schemas/Facts" }
        agentVersion: { type: string }
        online: { type: boolean }
        registeredAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
...